refactor: restructure filter plugin into modular architecture
